requirements.txt
veryfi>=0.1.8
pytest>=7.4.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0

